Most of PR feedback
