+++ src/WebJobs.Extensions.DurableTask/DurableOrchestrationStatus.cs
@@ -2,9 +2,6 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using DurableTask.Core;
 using Newtonsoft.Json.Linq;
 
 namespace Microsoft.Azure.WebJobs
