+++ src/WebJobs.Extensions.DurableTask/WebJobs.Extensions.DurableTask.csproj
@@ -61,7 +61,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Azure.DurableTask.Core" Version="2.1.3" />
-    <PackageReference Include="Microsoft.Azure.DurableTask.AzureStorage" Version="1.6.3" />
+    <PackageReference Include="Microsoft.Azure.DurableTask.AzureStorage" Version="1.6.4" />
     <PackageReference Include="Microsoft.Azure.Services.AppAuthentication" Version="1.0.3" />
   </ItemGroup>
