+++ test/FunctionsV1/WebJobs.Extensions.DurableTask.Tests.V1.csproj
@@ -33,14 +33,10 @@
     <Compile Include="..\Common\*.cs" />
   </ItemGroup>
 
-  <PropertyGroup Condition="true">
+  <PropertyGroup>
     <DefineConstants>FUNCTIONS_V1</DefineConstants>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Compile Remove="..\Common\DurableOptionsConfigurationTests.cs" />
-  </ItemGroup>
-  
   <ItemGroup>
     <ProjectReference Include="..\..\src\WebJobs.Extensions.DurableTask\WebJobs.Extensions.DurableTask.csproj" />
   </ItemGroup>
