+++ test/FunctionsV2/WebJobs.Extensions.DurableTask.Tests.V2.csproj
@@ -34,10 +34,6 @@
   <ItemGroup>
     <Compile Include="..\Common\*.cs" />
   </ItemGroup>
-
-  <ItemGroup>
-    <Compile Remove="..\Common\DurableOptionsConfigurationTests.cs" />
-  </ItemGroup>
   
   <ItemGroup>
     <ProjectReference Include="..\..\src\WebJobs.Extensions.DurableTask\WebJobs.Extensions.DurableTask.csproj" />
